+++ celeryconfig.py
@@ -1,2 +1,3 @@
 task_track_started = True
 task_ignore_result = False
+imports = ('gcpdac.celery_tasks')
